refactor: add service interface and implement service in serviceImpl
